% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{balloonPlot}
\alias{balloonPlot}
\title{draw balloon plot}
\usage{
balloonPlot(M, logTran, sqrtTran, main, namesx, namesy, margin, mult)
}
\arguments{
\item{M}{data}

\item{logTran}{do log transformation or not}

\item{sqrtTran}{do squareroot transformation or not}

\item{main}{name of this balloon plot}

\item{namesx}{column names}

\item{namesy}{row names}

\item{margin}{number of margin of this balloon plot}

\item{mult}{number of repetition of this balloon plot}
}
\value{
a balloon Plot
}
\description{
draw balloon plot
}

