% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{combind_fanwords}
\alias{combind_fanwords}
\title{combine two sparse matrix}
\usage{
combind_fanwords(A1, A2)
}
\arguments{
\item{A1}{sparse matrix 1}

\item{A2}{sparse matrix 2}
}
\value{
the large matrix which is the combination of A1 and A2
}
\description{
combine two sparse matrix
}
\keyword{combine}
\keyword{matrix}
\keyword{sparse}

