% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{create_fanwords}
\alias{create_fanwords}
\title{create the matrix of all users by words
users are the row subjects
words are the words post by users}
\usage{
create_fanwords(A)
}
\arguments{
\item{A}{adjacency matrix of users by posts}
}
\value{
the matrix of all users by words
}
\description{
create the matrix of all users by words
users are the row subjects
words are the words post by users
}
\keyword{adjacency}
\keyword{matrix}

