% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{createA}
\alias{createA}
\title{create adjacency matrix}
\usage{
createA(x, y)
}
\arguments{
\item{x}{the subjects to construct rows}

\item{y}{the subjects to construct columns}
}
\value{
the adjacency matrix of x and y
}
\description{
create adjacency matrix
}
\keyword{adjacency}
\keyword{matrix}

