% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{drop}
\alias{drop}
\title{take out the substring of the url of "https://www.facebook.com/123456789/posts/12345678"}
\usage{
drop(str)
}
\arguments{
\item{str}{the complete string of url}
}
\value{
the substring of url
}
\description{
take out the substring of the url of "https://www.facebook.com/123456789/posts/12345678"
}
\keyword{string}
\keyword{substring}

