% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{freqmatr}
\alias{freqmatr}
\title{find frequent words in TermDocumenMatrix}
\usage{
freqmatr(mytdm, rname)
}
\arguments{
\item{mytdm}{the TermDocumentMatrix}

\item{rname}{rownames of the matrix}
}
\value{
matrix of word frequency
}
\description{
find frequent words in TermDocumenMatrix
}

