% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{sv_bydate}
\alias{sv_bydate}
\title{do svd sorted by time}
\usage{
sv_bydate(v, datem, cc, npost, ncluster, ndate)
}
\arguments{
\item{v}{data to do SVD}

\item{datem}{time of data}

\item{cc}{membership vector of data (which post belongs to which candidate)}

\item{npost}{number of posts}

\item{ncluster}{number of clusters}

\item{ndate}{number of dates}
}
\description{
do svd sorted by time
}

