% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{svdmatrix}
\alias{svdmatrix}
\title{Do SVD to matrix with normalization}
\usage{
svdmatrix(A)
}
\arguments{
\item{A}{adjacency matrix}
}
\value{
top 20 singular values and singular vectors
}
\description{
Do SVD to matrix with normalization
}
\keyword{SVD}
\keyword{normalization}

