% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeClust.R
\name{textclean}
\alias{textclean}
\title{clean text
remove stopwords
remove numbers}
\usage{
textclean(x)
}
\arguments{
\item{x}{original text}
}
\value{
TermDocumentMatrix after cleaning
}
\description{
clean text
remove stopwords
remove numbers
}
\keyword{TermDocumentMatrix}
\keyword{clean}
\keyword{text}

